backend/requirements.txt
fastapi[standard]==0.121.2
psycopg2-binary==2.9.11
pydantic==2.12.4
python-dotenv==1.2.1
bcrypt==5.0.0
